+++ apps/ans/frontend/src/hooks/scan-proxy/useLookupAnsEntryByName.ts
@@ -12,7 +12,7 @@ const useLookupAnsEntryByName = (
   enabled: boolean = true,
   retryWhenNotFound: boolean = false,
   retry: number = 3
-): UseQueryResult<AnsEntry> => {
+): UseQueryResult<AnsEntry | null> => {
   const scanClient = useValidatorScanProxyClient();
 
   return useLookupAnsEntryByNameFromResponse(