+++ apps/common/frontend-typeface-termina/index.d.ts
@@ -1,5 +1,5 @@
 declare var common_typeface_termina: any;
 
-declare module "common-typeface-termina" {
+declare module "@lfdecentralizedtrust/splice-common-typeface-termina" {
   export = common_typeface_termina;
 }